gs backend: Use a clean file handle when copying using Reader
Under some circumstances gs.File.CopyToFile copies using f.tempFile as a
io.Reader. The problem with this is that if the cursor is not at
position 0, copying using io.Reader will not copy the whole file.

This change solves this problem by using a clean reader when copying
using io.Reader so that the copy is not affected by the reader cursor
position.