gs backend: Return an error when trying to copy/move a file with non 0 cursor position. (#80)
Under some circumstances gs.File.CopyToFile copies using f.tempFile as a
io.Reader. The problem with this is that if the cursor is not at
position 0, copying using io.Reader will not copy the whole file.

This change adds a check to gs.File.CopyToFile that will return an
error if the method is called on a file with non 0 cursor position.