Deprecated TouchCopy and using TouchCopyBuffered instead in backend.
TouchCopyBuffered allows for custom file buffer size to be set in code where TouchCopy previously was used, and it sets default buffer size to 256MB vs 32 KB found in go io.Copy. A 256MB file buffer results in 46% faster copies on larger files ~1GB+ with minimal impact on smaller file transfers.

Added tests in utils_test.go for TouchCopyBuffered

Added tests in GCP, Azure, S3, SFTP, and OS to test buffered copy mods in respective file_test.go for each provider.

Fixed warning on reader.Close() in code analysis for an uncaught/potential hidden exception in a deferred function in GCP tests.

Pending Need To refactor handling of options to be consistent and if interfaces are used must handle pointers and values. This scope will be handled in separate PR.