Updated changelog ---- Merged SFTP From Master ---- Updated spacing on utils.md to be tabs vs. spaces ---- Updated internal calls to io.Copy to use io.CopyBuffer and utils.TouchCopyMinBufferSize ---- Updated per PR comments
----
Deprecated TouchCopy and using TouchCopyBuffered instead in backend.

TouchCopyBuffered allows for custom file buffer size to be set in code where TouchCopy previously was used, and it sets default buffer size to 256KB vs 32 KB found in go io.Copy. A 256MB file buffer results in 46% faster copies on larger files ~1GB+ with minimal impact on smaller file transfers.

WIP1

Fixed warning on reader.Close() in code analysis for an uncaught/potential hidden exception in a deferred function in GCP tests (Not Part of Code Ticket Scope)

Fixed Lint Warnings

Pending Need To refactor handling of options to be consistent and if interfaces are used must handle pointers and values. This scope will be handled in separate PR.